requirements.txt
PyYAML==6.0.2
Requests==2.32.4